# _replit_import_20260506_153127/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ConnectSphere API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/status:
    get:
      operationId: getSubscriptionStatus
      tags: [subscriptions]
      summary: Get current user subscription status
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
  /subscriptions/checkout:
    post:
      operationId: createCheckoutSession
      tags: [subscriptions]
      summary: Create a Stripe checkout session for premium subscription
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
  /subscriptions/portal:
    post:
      operationId: createCustomerPortal
      tags: [subscriptions]
      summary: Create a Stripe customer portal session
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSessionResponse"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL for profile photos
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
  /storage/objects/{path}:
    get:
      operationId: getObject
      tags: [storage]
      summary: Retrieve a stored object
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object file
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    SubscriptionStatus:
      type: object
      properties:
        isPremium:
          type: boolean
        plan:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        currentPeriodEnd:
          type: string
          nullable: true
        cancelAtPeriodEnd:
          type: boolean
      required:
        - isPremium
    CreateCheckoutBody:
      type: object
      properties:
        priceId:
          type: string
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - priceId
        - successUrl
        - cancelUrl
    CheckoutSessionResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    PortalSessionResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...
    RequestUploadUrlBody:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
      required:
        - filename
        - contentType
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        objectPath:
          type: string
      required:
        - uploadUrl
        - objectPath
...